+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2014-05-31
+## [0.0.2] - 2022-05-26
 ### Added
-- runSpec command
-- initial tspec language support
+- Document formatting.
+
+## [0.0.1] - 2022-05-26
+### Added
+- runSpec command.
+- Initial tspec language support.
 
 [Unreleased]: https://github.com/tspec/vscode-tspec-runner/compare/v0.0.1...HEAD
+[0.0.2]: https://github.com/tspec/vscode-tspec-runner/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/tspec/vscode-tspec-runner/releases/tag/v0.0.1